refactor: consolidate context directories into /src/contexts
Move SubGenieContext.tsx from /src/context to /src/contexts to have
all context providers in a single directory. Update imports in App.tsx
and useSubGenieWidget.ts.
